fixes cache and commands

+++ src/main/java/com/agmadnasfelguc/walgreensreplica/user/service/command/LogoutCommand.java
@@ -1,11 +1,13 @@
 package com.agmadnasfelguc.walgreensreplica.user.service.command;
 
 import com.agmadnasfelguc.walgreensreplica.user.cache.SessionCache;
+import com.agmadnasfelguc.walgreensreplica.user.service.Utils.JwtUtil;
 import com.agmadnasfelguc.walgreensreplica.user.service.command.Command;
 import com.agmadnasfelguc.walgreensreplica.user.service.command.helpers.ResponseFormulator;
 import com.agmadnasfelguc.walgreensreplica.user.service.response.ResponseState;
 import com.agmadnasfelguc.walgreensreplica.user.service.response.ResponseStatus;
 import lombok.*;
+import lombok.extern.slf4j.Slf4j;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -15,22 +17,29 @@ import java.util.Map;
 
 @EqualsAndHashCode(callSuper = true)
 @Service
-@Data
+@Slf4j
 public class LogoutCommand extends Command {
+    @Setter
     private String sessionId;
 
     @Autowired
     private SessionCache sessionCache;
-    Logger logger = LoggerFactory.getLogger(LogoutCommand.class);
     @Override
     public void execute() {
         try{
-            sessionCache.deleteCompleteSession(sessionId);
-            this.setState(new ResponseStatus(ResponseState.Success, "Logged Out"));
+            String userId = JwtUtil.getUserIdFromToken(sessionId);
+            if(userId == null){
+                this.setState(new ResponseStatus(ResponseState.Failure, "Invalid Session"));
+            }
+            else{
+                JwtUtil.blacklistToken(sessionId);
+                sessionCache.deleteCompleteSession(sessionId);
+                this.setState(new ResponseStatus(ResponseState.Success, "Logged Out"));
+            }
         }catch (Exception e){
             this.setState(new ResponseStatus(ResponseState.Failure, e.getMessage()));
         }
-        ResponseFormulator.formulateResponse(logger, this.getState(), this.getReplyTopic(), this.getCorrelationId(), this.getUserRequests(), null);
+        ResponseFormulator.formulateResponse(log, this.getState(), this.getReplyTopic(), this.getCorrelationId(), this.getUserRequests(), null);
 
     }
